feat: files "new file" menu generate image action

+++ lib/Listener/TaskSuccessfulListener.php
@@ -10,11 +10,14 @@ namespace OCA\Assistant\Listener;
 use OCA\Assistant\AppInfo\Application;
 use OCA\Assistant\Db\TaskNotificationMapper;
 use OCA\Assistant\Event\BeforeAssistantNotificationEvent;
+use OCA\Assistant\Service\AssistantService;
 use OCA\Assistant\Service\NotificationService;
 use OCP\EventDispatcher\Event;
 use OCP\EventDispatcher\IEventDispatcher;
 use OCP\EventDispatcher\IEventListener;
+use OCP\IURLGenerator;
 use OCP\TaskProcessing\Events\TaskSuccessfulEvent;
+use Psr\Log\LoggerInterface;
 
 /**
  * @template-implements IEventListener<Event>
@@ -25,6 +28,9 @@ class TaskSuccessfulListener implements IEventListener {
 		private TaskNotificationMapper $taskNotificationMapper,
 		private NotificationService $notificationService,
 		private IEventDispatcher $eventDispatcher,
+		private AssistantService $assistantService,
+		private LoggerInterface $logger,
+		private IUrlGenerator $url,
 	) {
 	}
 
@@ -38,7 +44,11 @@ class TaskSuccessfulListener implements IEventListener {
 			return;
 		}
 
-		if ($this->taskNotificationMapper->getByTaskId($task->getId()) === null) {
+		$customIdPattern = '/^new-image-file:(\d+)$/';
+		$hasTargetDirectory = preg_match($customIdPattern, $task->getCustomId(), $matches) === 1;
+
+		// For tasks with customId "new-image-file:<directoryIdNumber>" we always send a notification
+		if ($this->taskNotificationMapper->getByTaskId($task->getId()) === null && !$hasTargetDirectory) {
 			return;
 		}
 
@@ -57,7 +67,31 @@ class TaskSuccessfulListener implements IEventListener {
 			$notificationActionLabel = $beforeAssistantNotificationEvent->getNotificationActionLabel();
 		}
 
+		if ($hasTargetDirectory) {
+			$directoryId = (int)$matches[1];
+			$fileId = (int) $task->getOutput()['images'][0];
+			try {
+				$file = $this->assistantService->saveNewFileMenuActionFile($task->getUserId(), $task->getId(), $fileId, $directoryId);
+				$notificationTarget = $this->url->linkToRouteAbsolute(
+					'files.viewcontroller.showFile',
+					[
+						'fileid' => $file->getId(),
+						'opendetails' => 'true',
+						'openfile' => 'false',
+					],
+				);
+			} catch (\Exception $e) {
+				$this->logger->error('TaskSuccessfulListener: Failed to save new file menu action file.', [
+					'task' => $task->jsonSerialize(),
+					'exception' => $e,
+				]);
+			}
+		}
+
 		$this->notificationService->sendNotification($task, $notificationTarget, $notificationActionLabel);
-		$this->taskNotificationMapper->deleteByTaskId($task->getId());
+
+		if (!$hasTargetDirectory) {
+			$this->taskNotificationMapper->deleteByTaskId($task->getId());
+		}
 	}
 }
